Migrated from data components to attributes
Migrated from using data components to attributes for weapons.

Added base melee weapon class.
Simplified existing magic weapon class.
Added new sword.

+++ src/main/java/com/duubl/via_arcana/magic/IMana.java
@@ -7,10 +7,10 @@ public interface IMana {
 
     void addMana(int amount);
 
-    boolean consumeMana(int amount);
+    boolean consumeMana(double amount);
 
     int getMaxMana();
 
-    float getRegenRate();
-    void setRegenRate(float rate);
+    double getRegenRate();
+    void setRegenRate(double rate);
 }
